@@ -1,26 +1,29 @@
 name: zakupyapp
 description: Lista zakupowa
-version: 1.5.3  # updated 19/01/2024
+version: 1.6.0  # updated 03/08/2024
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^2.15.1
-  firebase_database: ^10.2.5
-  shared_preferences: ^2.2.1
-  package_info_plus: ^5.0.1
-  test: ^1.24.3
+  firebase_core: ^3.3.0
+  shared_preferences: ^2.3.1
+  package_info_plus: ^8.0.2
+  test: ^1.25.7
   url_launcher: ^6.1.14
   provider: ^6.0.5
   diffutil_sliverlist: ^0.5.1
-  http: ^1.2.0
+  http: ^1.2.2
+  firebase_auth: ^5.1.4
+  google_sign_in: ^6.2.1
+  cloud_firestore: ^5.2.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^4.0.0
 
 flutter:
   uses-material-design: true
