@@ -1,7 +1,6 @@
 class Constants {
-  static const String FAMILY_STORE_APP_URL =
+  static const String familyStoreAppUrl =
       'https://family-store.vercel.app/apps/14';
-  static const String FAMILY_STORE_ENDPOINT =
+  static const String familyStoreEndpoint =
       'https://family-store.vercel.app/api/app-version/14';
-  static const String DEFAULT_USERNAME = 'Użytkownik bez nazwy';
 }
